Check alternating paths using iterative DFS in to_vertex_cover.
This PR fixes #2384.

Reimplement iteratively the inner function `_alternative_dfs` in
`_is_connected_by_alternating_path`. This function checks if nodes
in the graph are linked to target nodes by alternating paths, that
is, paths that alternate between matched and unmatched edges.

This PR also delegates to `_connected_by_alternating_paths`
computing the matched and unmatched edges because these are the same
for a given maximum matching and we do not need to recompute them
each time that we check if a node is linked to targets by an
alternating path.

I also changed the implementation of the computation of matched and
unmatched edges in order to avoid assuming orderable nodes.

Added a test with the example given in #2384 and another test with
unorderable nodes.