Improve bipartite documentation.
This PR addresses #2400

Improve bipartite documentation by clarifying that the use of a node
attribute named `bipartite` with values 0 and 1 in order to
distinguish bipartite node sets is only a convention which is not
enforced in the source code. Also state explicitly that bipartite
functions do not check that the input graph is actually bipartite
nor than the container with a bipartite node set contains all nodes
from that bipartite node set.

Added links form all relevant function docstring to the module
level documentation where all this is explained.

Also added links to some custom exceptions to the generated
documentation that were missing.