Install gdal after creating the conda environment.
This workaround still does not install gdal for python 3.6 but
it doesn't kill the build, and it works for all other python
versions. It seems that the probelm is that conda for python 3.6
installs the latest version of gdal (2.1.3) which is not compatible
with the package libgdal-dev provided by Travis version of Ubuntu.