added documentation for write_gexf viz attrs
I added a sentence describing that networkx does indeed write
attributes like visualization parameters, "viz", correctly according
to the GEXF schema.

I added an example setting the size color and layout positions of a
node which was tested to work in Gephi 8.0.2.

There was a few questions on stackoverflow and some promoting
work-around behavior because it was not evident in the documentation.

I will note that beyond viz there are other behaviors that are not
documented and the user would not know they are available without
actually reading through the code.