Remove stale reference to pngmath (#2485)
* Remove stale reference to pngmath, fixes #2340

Also see commit ffa3e5deb837

* Fix import warning from sphinx-build:

    WARNING: [autosummary] failed to import u'networkx.algorithms.community.generators.LFR_benchmark_graph': no module named networkx.algorithms.community.generators.LFR_benchmark_graph

The module is named `community_generators` not `generators`.

* Remove all reference to old and unused indexsidebar.html

* Remove old references to the Trac site including remoing unused
extension code

* Update copyright years

* Remove mathjax path

It isn't necessary to specify this (unless you want to use
a local path or an alternative URL other than the cdnjs
Content Delivery Network.  Also the headcoded URL was retired
in April, 2017 according to
http://docs.mathjax.org/en/latest/start.html.

* Clean up requirements

The project is not using bibtex and it is no longer necessary to use
Sphinx stable branch via GitHub.