rebase: Moves is_partition to community.utils.
This makes the function `is_partition` public, and ensures that it only
appears in one place. This commit also changes the `community` package
so that it behaves like the `bipartite` package, and various others: its
functions are not exported to the top-level `networkx` namespace; the
functions must be accessed through `networkx.algorithms.community`.

Finally, this commit also simplifies some of the community tests and
makes them more readable.