Adds structural hole measures (#2378)
* Adds functions computing structural hole measures

This commit creates a new module, `networkx.algorithms.structuralholes`,
that provides functions for measuring structural holes in a graph.

* Fix, improve, and speed-up structural holes measures.

This PR supersedes #1535, #429, and #192.

This commit fixes the implementation of constraint and effective size in
PR #1535. That implementation was using `nx.all_neighbors` to iterate
over all predecessors and successors of nodes. The problem was that if a
neighbor is both a predecessor and a successor is appears twice in
`nx.all_neighbors` and thus it was double counted.

Regarding constraint, PR #1535 was following the formula of constraint
posted at a very visible web page [1], which is wrong as it says that
the local constraint has to be calculated only on neighbors that are
both predecessors and successors of the focal node. Burt's paper [2]
at page 362 states that "I measure the constraint on each manager with
respect to the immediate network of discussion partners, composed of
anyone the manager cited as a discussion partner and anyone who cited
the manager.", which means that it should be computed for all neighbors
that are either successors or predecessors.

Regarding effective size, I implemented the original formula proposed by
Burt in [3], which was different that the one used in #1535. For the
case of unweighted and undirected graphs I implemented the simplified
formula proposed by Borgatti [4] which is faster.

After these changes the tests posted in #192 now pass. These test were
checked using UCINET, I rechecked them using IGraph and also computed them
by hand to make sure that they were right as I don't have access to
UCINET and effective size is not implemented in IGraph. I also added more
tests to cover all cases applicable (weighted, unweighted, directed,
and undirected).

I changed the API of the functions to compute the measures on all nodes
of the graph and accept an arbitrary edge weight name to compute the
weighted versions, following the API of other centrality measures in
NetworkX.

After profiling these measures, I sped-up the running time by
reimplementing the inner function `mutual_weight` as a succession of
`try` and `except` instead of the original generator expression, which
is ~40% faster. This inner function is called millions of times for not
so large graphs (thousands of nodes) and thus the speed up is
significant. On the other hand, I also removed the decorator
`not_implemented_for` of this function, which according to profiling
accounted for 50% of the total running time of the measures.

[1] http://www.centiserver.org/?q1=centrality&q2=Burts_constraint
[2] Burt, R. (2004). "Structural holes and good ideas". American Journal
of Sociology (110): 349–399.
[3] Burt, Ronald S. (1995). Structural Holes: The Social Structure of
Competition. Cambridge: Harvard University Press.
[4] Borgatti, Stephen (1997). "Structural Holes: Unpacking Burt's
Redundancy Measures". Connections. INSNA.