Graphml: consistent I/O for key and data attributes of edges in multi-graphs (#2559)
* Graphml: read and write key and data attributes of edges in multi-graphs.

Before this commit, key and attribute of multi edges were both stored as `data` xml elements of graphml. This caused difference in edges of multigraphs with key and data attributes after consequent write and reads of graphml. `test_preserve_multi_edge_data` was added to `test_graphml.py` to show this inconsistency.
This commit changes graphml reader and writer behavior to:

* read and write `key` of networkx edges only in `id` attribute of `<edge id="...">` elements of graphml.
* read and write `data` of networkx edges only in `<data></data>` elements of graphml.

Only in non-multigraphs, `id` attribute of `<edge id="...">` elements will be read into `data` of networkx edges, because we have no `key` for edges.

* Fix pep8 errors for graphml