Next attempt to meld graphviews with base classes (#2593)
* Update code to prepare for melding graphviews

* Meld graphviews into graph classes

* Cleanup subgraph calling sign. and remove duplicate code

* Add some tests for raising exceptions

* update edge_kcomponents to avoid readonly views.

* Add root_graph attribute and tests

Update tests for root_graph as well as fresh_copy.
I left fresh_copy as an attribute even with root_graph
because a view might switch the data structure of the
view from directed to undirected. Going to the root_graph.__class__
may not give you what you need to create a graph like the view.
Fresh_copy gives a null graph with the directed/multi type of that
view or graph.