Rewrite fiedler tracemin_pcg
The Fiedler vector pcg method in rare (but sometimes simple) cases
projected away too much energy from b in Ax=b so that x was within
roundoff error and the solver never finished (hung). Despite efforts
to restore the energy, add noise in orthogonal directions and project
either more or less often, the system would hang on unlucky initial
conditions. (Though I did get it to only hang 5 times out of 1 million
seeds.)

In the end I took a hint from a comment line that the PCG method was
using a different but mathematically equivalent method to that of
Cholesky and LU. Rewriting the PCG method to use the same mathematical
statement as the Chol and LU methods did the trick. Many fewer
projections might save time, certainly simpler code, and the
vectors are not close to round-off size. Trials with simple graph
and initial conditions from seeds 1:1000000 show no signs of hanging.

I believe we can now remove the restriction in the tests that the seed
be the same each time we test.  Is there a reason to keep the seeds
the same other than avoiding an infinite loop?