Sparse fruchterman needs t scaled by domain size (#2868)
Unlike the regular version of fruchterman reingold, the sparse version doesn't calculate temperature proportionality to domain size. I propose to make the exact same calculation for t in the sparse version, too.
I noticed this when drawing a graph with a lot of concentrated points with faraway, fixed position (edge length in order of 1-3 and position > 300000, 300000 ) , because a couple of non-fixed nodes refused to get near the others.