Fixed strongly_connected_components [Issue 2831]
Since we are using iterative method we spend too much time on
rechecking graph vertices that are already have been checked.
Removing vertices that already detected as strongly connected from the
copy of the graph dramatically speeds up this method.
Also using dict of lists and removing checked elements instead of
checking them in a loop increased the
speed.