Fix UnionFind set extraction
Paths are not guaranteed to be fully pruned by the current
implementation if subtrees are unioned, e.g.:
> uf = UnionFind()
> uf.union(1, 2)
> uf.union(3, 4)
> uf.union(4, 5)
> uf.union(1, 5)

In the current implementation, parents[1]=3 and parents[2]=1.
Thus, the mere call of networkx.utils.groups(parents) will yield a
wrong result:
[set([2]), set([1, 3, 4, 5])]

This patch fixes this behavior by simply doing a "find" operation
(which, in turn, does full path pruning) on every key before calling
groups.