Add barycenter algorithm (#2939)
* WIP add barycenter algorithm

* attr parameter allows saving barycentricity value for each node

* Fix error in test that didn't read updated calculation

Should have included this in fdd22f1002

* Tests don't depend on NumPy, and improve docs

* More doc clean up

* Take advantage of new randomness API in NetworkX 2.2

* barycenter now returns nodes rather than subgraph

This unifies barycenter's signature with center and periphery. See also
https://github.com/networkx/networkx/pull/2939#issuecomment-407757160

* Cross reference docs for center, barycenter, and periphery

* Clarify and simplify doc strings

Also clarify name of a test utility function

* Alphabetize imports

* PEP8: Indent test module with spaces

* Add sp keyword argument to barycenter

This allows barycenter to use pre-computed shortest path lengths.

* Make test more specific

* Doc string style

* PEP8: line lengths

* Weights belong on the edges, not vertices

The test is checking that you can't use the weight and sp arguments at the same
time. For it to be even plausible that we're using the weight argument, the
graph needs weights: *edge* weights. Not node weights as the previous version of
the test had.

* Move TestBarycenter from test_barycenter to test_distance_measures

I don't know the nose package, so I was afraid of trying to develop the tests in
the existing test_distance_measures module. However nose seems to play nicely
with the standard unittest package, so we're all good.

* Clean up tests in test_distance_measures.py