Split is_at_free method
The previous method had an extra parameter, `certificate`, that could be used
to decide whether the output should be a boolean value or a tuple with a
boolean value and a certificate. This has now been split into two functions:
`is_at_free` and `find_asteroidal_triple`, for more intuitive behaviour.