AT-free graph recognition (#3377)
* Included asteroidal module.

* Initial commit for asteroidal module.

The asteroidal module implements a function which checks whether
a given graph is AT-free or not.

* Tests for asteroidal.py

The tests include checks for small, well-known cases of
AT-free and non-AT-free graphs.

* Updated runtime for recognition algorithm in docstring.

* Included "r" to start docstring to enable LaTeX codes.

* Included .rst file for asteroidal module

* Fixed not implemented for decorator statements

If multiple types are inputed into a decorator this
works as an AND condition, but we want an OR condition.

* Refactoring of unnecessary work

The helper function in is_at_free now consists of a return
statement directly, which saves having to evaluate boolean
statements unneccesarily. The for loop in the component_strucutre
function now omits an unecessary generator definition.

* Updated docstring to reflect optional output and runtime.

The docstring now includes a more precise description of the
worst-case runtime. The optional output parameter `certificate`
is more thoroughly documented.

* Change optional output to a tuple including a bool

* Refactor tests to be shorter and more readable

* Included tests with certificate option enabled.

The certificate option outputs a certificate, if any exists,
which proves that a graph is not AT-free.

* Fix bug due to typo in tests

An 's' had snuck its way into an assert_equal statement,
and I didn't check properly before pushing.

* docstring for is_at_free re-structured and re-formulated.

* Split is_at_free method

The previous method had an extra parameter, `certificate`, that could be used
to decide whether the output should be a boolean value or a tuple with a
boolean value and a certificate. This has now been split into two functions:
`is_at_free` and `find_asteroidal_triple`, for more intuitive behaviour.

* Changed `is_at_free` to find_asteroidal_triple

The new version returns an asteroidal triple if one exists
and returns None if none exists. This method is used in the
new `is_at_free` method.

* Updated docstrings

Docstrings updated to reflect changed functionality and
new structure.

* Refactor `find_asteroidal_triple`

The previous version defined a function to check whether a
triple of vertices is an asteroidal triple. The current
version uses this function inline instead, to avoid many
function calls in order to increase performance.

* Refactored return statement in `is_at_free`

Removed the else block according to Chromium style guide.

* feat: expose `find_asteroidal_triple`

The find_asteroidal_triple is also imported when the asteroidal
module is imported.

* style: refactor is_at_free

The function can be written as a readable oneliner instead
of using an if else block.

* docs: asteroidal: updated module docstring

Rewrote module docstring to be more precise.

* chore: include find_asteroidal_triple in reference

The find_asteroidal_triple is now exposed when importing
the asteroidal module so the method should also be included
in the asteroidal.rst for auto-generation of documentation.

* Added myself, Haakon H. Rød, as a contributor

* chore: removed redundant tests

Some tests are no longer valid after changing the output
of `is_at_free` to purely boolean.