Remove boiler plate from top of modules
The copyright and author stuff is not necessary, out-of-date,
and inconsistent.  It takes up visual space and is a pain to
police everyone doing the same thing on the top of the module.

Git handles authorship in a comprehensive and authoritative way.
The LICENSE.txt file applies to all project code.