Remove completeness condition from minimum weight full matching
When first implementing `minimum_weight_full_matching`, we were still
only requiring a SciPy version of 1.1.0, which meant that we had to
require that the bipartite graph was complete. This requirement was
removed in SciPy 1.4, and now (with 8d8aa86eede1bccc160af73e823492357196b341)
that NetworkX specifies a minimum SciPy requirement of 1.4, the
requirement can be removed here as well.