API: Add  kwarg to view_pygraphviz.
Adds a new keyword argument to view_pygraphviz to toggle whether or not the
drawn graph will actually be displayed.

Currently, there are tests in test_agraph the result in graphs being displayed
via 3rd party viewers (platform dependent) but that don't actually test
whether the rendered graphs are correct.