Don't import other people's version.py
fix #4288

When a repository calls import networkx it will
trigger a sequence of imports that will cause to
look for a version.py in the current working
directory. Because it is common for a repository to contain
a version.py, this will end up being executed.

The change guarantees that this library version.py is called.

Arguably, the line of code should not being called at all when
importing networkx.