Switch all() generator to for loop and check edge is valid
This commit fixes an issue where in some scenarios an invalid edge (it
looks like from self loops) is being passed in in the matching set. To
correctly hand this, this commit moves away from a generator and just
uses a for loop to check that the edge has 2 nodes and that it's a valid
edge in the graph.