Verify edges are valid in is_matching() (#4638)
* Verify edges are valid in is_matching()

Previously the is_matching() function was never checking the edges in
the provided matching against the graph, it would just validate there
were no shared endpoints. The graph object passed in to the function was
never used. This commit updates the function to check that all the edges
in the matching are valid and present in the provided graph before
checking if the are no shared endpoints.

Co-authored-by: Dan Schult <dschult@colgate.edu>