https links for conference.scipy.org documents
The server reset from https://github.com/networkx/networkx/issues/5433 fixed the resolution issues, but both Firefox and Chromium refuse to play nice with the .`bib` file on this machine, apparently because they're concerned with the insecurity of the `http:` target. Both are happy to download the .bib from `https:...`. I've changed the PDF and .bib links to `https`, which they accept.