fix bug for directed graphs
if you read the original document for directed louvain (https://hal.archives-ouvertes.fr/hal-01231784/document), you'll see that in section 2, the equation for gain in the undirected case (delta Q) has a factor of 2 divided into both sides of the equation, whereas the directed case has that factor missing from both sides. In the previous code here, the factor was missing from only the first fraction, so I also "remove" it here to balance things out. Equivalently, this makes "resolution" twice as effective as before, thereby matching its strength in the undirected case