Setup vf2pp files for expansion to subgraph problems
Notes:

- Reordered the intro of the main function vf2pp_all_isomorphisms.
  - Replace `return False` (which doesn't do much inside a generator)
    with `return "helpful string"`. The string is only seen if you manually
    use `next` on the generator functuion. But that can be helpful for debugging.
  - Moved the degree dict creation to `_initialize_parameters`. Moved this call
    to before where the degrees are used.
  - Inline the function `precheck_label_properties` and make more readable.
- added G1_degree and G2_degree degree dicts to the graph parameters namedtuple.
  Removed the degree dict input parameter to some helper functions.
- Reworded, unified language and corrected the main module doc_strings.
- changed the name of utils.labels_different to utils.labels_many.
- adjusted the tests accordingly.

Adding the features for "sub" and "mono" problem types should be easier after
these changes.

For the future:

The namedtuples are very hard to adjust throughout the code. It would be good
to have a central place instead of unpacking the namedtuple in each function.
But maybe there isn't a better way.

I think the Directed version of the helper functions can be eliminated with
code that checks during the function. I also believe the namedtuples could
be better used here, perhaps being different between graph and directed graph.

The multiedge calculation should be moved to _initialize_parameters
to avoid recalculation for MultiGraphs.