Refix minimum_cycle_basis and scipy.sparse conversions and add tests (#6789)
* update tests to check basis independence

* fix new minimum_cycle_basis

* fix subtle type bug when converting from scipy sparse

The nodes were being added as python int (via range(n))
But the edges were being added as np.int32 from csr_matrix.indices
The nodes are equal to the nodes used when forming the edges, so no
new nodes were added and it all seems to work!
But np.int32 raises a ValueError when comparing to tuples while python
int shows `False`.  This leads to a very cryptic error message if the
graph later adds any nodes that are tuples.