Be forgiving of iteration order in `test_write_network_text_circular_ladder_graph`
This is helpful for backends that implement `circular_ladder_graph`.
Currently, the backend testing infrastructure in NetworkX can call backend
graph generators such as `circular_ladder_graph`, and it converts the backend
Graph to a networkx Graph. This is a great way to test backend generators, but
sometimes iteration order is different.

Alternatively, the testing infrastructure could call *both* the original
generator and the backend generator, check that the graphs are equal, then
use the original graph. This approach would avoid other test failures that
result from difference in iteration order.