Compare graphs for generator functions when running tests with backend
This is an alternative to #7063.

The backend testing infrastructure can call backend graph generator functions
such as `circular_ladder_graph`, and it converts the backend Graph to a networkx
Graph. This is a great way to test backend generators! However, sometimes iteration
order is different when using the converted backend graph. So, our solution here
is to *also* call the original networkx function when the return type is a Graph,
compare results, then use the result from networkx to allow tests to pass.
We use a series of assertions to help backend implementers identify differences.