Formatting issues have now been rectified using pre-commit. Please note that test_coloring.py had been altered because (a) the new version allows some paramter combinations that were previously disallowed, and (b) additional parameter options are also available for greedy_coloring and equitable_coloring. I hope that this is OK
At a higher level - the reason why this constraint exists is that these functions are already in use by users for all sorts of applications. It's important to ensure that any changes that go into the library don't silently yank the rug out from under them and change the behavior of existing code, at least not without some form of forewarning.