Try to Remove autogenerate requirements files.
Where possible in the github action use `.[...]` syntax

With the exeption  of lint that was not installing networkx and now
does I hope that this should be identical

It does sometime merge a few pip command into a single one.

Note that some of those files are not autogenerated, so leave them.

One of the limitation is that the instructions to intall with conda
(`conda install -c conda-forge --file requirements/extra.txt`) don't
make sens anymore.

This would replace #7483