DOC: Rm redundant module from autosummary. (#7599)
Autosummary started raising warnings in the sphinx 7.4.X
series related to multiple specifications of the parent module
in autosummary directives.

There were some such instances in the tutorial. In this case the
 was unnecessary because the
was specified at the top of the tutorial.

These changes preserve correct linking and get rid of the sphinx warnings.