Update `_raise_on_directed` to work with `create_using` pos arg
This decorator is only used by the "small" graph generators.
We could update this to use `argmap`, but probably not worth the effort
since all functions have create_using as the first positional argument.

If we were creating these functions today, we might make create_using
keyword-only, but that would be a breaking change.