Fix bug in modular decomposition
The original implementation was limited to input graphs with integer
nodes and, thus, an "isinstance(int)" check was enough to determine
whether a node in the MD-tree belonged to the original graph (i.e., a
leaf in the MD-tree). The NetworkX port, however, should support input
graphs with arbitrary node types.

Change all "isinstance(int)" checks to "not isinstance(Node)" checks.