Fixes #6873 - MEO/PEO handling in chordal.py
We extract the MCS-M triangulation logic into minimal_elimination_order and ensure that MCS-M is run even if the input graph is already chordal. We keep complete_to_chordal_graph for backwards compatibility, simply returning early if the input graph is chordal (but with the caveat that the alpha output will be all zeros, which is what motivated this issue/PR in the first place) and calling minimal_elimination_order otherwise. We also create a new function perfect_elimination_order which assumes that the input graph is already chordal, raising an error if it is not.

One more step before merging is to change the returned 'alpha' variable by minimal_elimination_order and perfect_elimination_order to a list rather than a dict. We intend to keep the current behavior in complete_to_chordal_graph for API compatibility, however.