Use type checks in `generators/lattice.py` for Py3.14 compat
Replace the `TypeError` catching in `networkx/generators/lattice.py`
with explicit type checks to make the code more reliable and fix it for
Python 3.14.  Catching the exception immediately does not work
in the second instance, because the code is constructing a generator,
and apparently Python 3.14 does not evaluate the `p in periodic`
expression until the generator is actually iterated over.  Given that
the function expects either an iterable or a `bool`, explicitly checking
for `bool` should both be more readable and more reliable.

The alternative would be to replace the generator with a list
comprehension that would be evaluated immediately.  However,
the explicit type check seems to be a cleaner solution to the problem.