DOC: (`cutoff` edges) -> (`cutoff + 1` nodes)
Also adds an example of `cutoff` to the `predecessor()` docstring,
while we're at it. There are some complications w/r/t the cutoff
parameter of that function (see comments on the PR from last year),
so this PR is not modifying anything in the docs there beyond
cosmetic / example changes.