Clarify the meaning of the cutoff parameter in some path-finding functions (#7487)
* DOC: clarify cutoff wrt edges for simple path docs

* DOC: clarify cutoff for shortest path functions

* BUG: allow cutoff=0 for predecessor()

* MNT: Revert "cutoff is None" check in predecessor

* DOC: say "`cutoff`" instead of "cutoff"

matches all_pairs_shortest_path(), and looks nice in the online
docs

* DOC: length clarifications for all_pairs*()

* DOC: use ticks for relevant simple paths funcs

* DOC: cutoff ex. - all_pairs_shortest_path_length

* DOC: cutoff examples for single_*_shortest_path

also small wording change in the all_pairs_shortest_path_length
example

* DOC: finish adding cutoff examples for the 8 funcs

* DOC: adjust all_simple*paths() cutoff descs

- don't say "edges", as discussed
- use a better example for the all_simple_edge_paths() cutoff desc

* DOC: (`cutoff` edges) -> (`cutoff + 1` nodes)

Also adds an example of `cutoff` to the `predecessor()` docstring,
while we're at it. There are some complications w/r/t the cutoff
parameter of that function (see comments on the PR from last year),
so this PR is not modifying anything in the docs there beyond
cosmetic / example changes.

* DOC: use double backticks around "cutoff + 1"

* Minor tweaks for numpydoc style examples.

* tweak wording in a couple spots

---------

Co-authored-by: Ross Barnowski <rossbar@caltech.edu>
Co-authored-by: Dan Schult <dschult@colgate.edu>